[3, 4단계 방탈출 예약 대기] 레모네(이지현) 미션 제출합니다. (#159)
* docs(README): ERD 수정

* rename(...): infra, vo, config 패키지 생성

* rename(... TimeService): Time 네이밍을 모두 ReservationTime으로 변경

* chore(application.yml): ddl 생성 create-drop으로 변경

* refactor(... MemberController): 사용자별 예약 목록 조회 endpoint /members/.. -> /member/.. 로 변경

* refactor(MemberIntegrationTest): 어떤 회원의 어떤 예약 내역을 가져올 것인지 메서드를 추가해 메서드명을 통해 명시

* style(RoomEscapeApplication): 공백 제거

* docs(README): 3단계 요구사항 추가

* feat(reservation.html, user-reservation.js) : 요구사항에 맞게 클라이언트 코드 추가

* rename(... registration 패키지): reservation 패키지를 registration 패키지 안으로 이동, 그 외에도 가시성이 좋게 패키지 분리 수

* feat(WaitingRequest, WaitingResponse): 예약 대기용 요청, 응답 dto 생성

* feat(WaitingService, WaitingRepository): 예약 대기 레포지토리 & 서비스 생성

* feat(ReservationStatus): 예약 상태 enum에 예약 대기 상수 추가

* feat(Waiting): 예약 대기 테이블 추가

* feat(user-reservation.js): 예약 대기 생성 요청 api 호출

* feat(WaitingController): 예약 대기 api 추가

* feat(WaitingExceptionCode): Waiting용 예외 상수 추가

* feat(WaitingWithRank): 예약 대기 순서 반환을 위한 dto 추가

* remove(ReservationWaitingResponse): 사용하지 않는 dto 제거

* docs(README): 예약 대기 api 완성

* refactor(resources): 예약과 예약 대기를 모두 보여주는 api는 reservation -> registration으로 uri 변경

* feat(MemberController): 사용자의 예약 목록을 보여주는 api 주소를 /member/reservations -> /member/registration 으로 변경 후 반환값에 예약 대기 내역 추가

* refactor(ReservationResponse): 정적 팩토리 메서드명 간결하게 수정

* feat(... WaitingRepository): Waiting 내역과 내 앞 대기열 수 반환하는 레포지토리 메서드 추가

* style(WaitingController): 주석 제거

* docs(README): 내 예약 목록 조회 시 예약 대기 목록이 보이는 기능 추가

* feat(... WaitingController): 예약 대기 삭제 기능 추가

* refactor(Reservation): 예약 시간 필드 변수 이름 time -> reservationTime 으로 변경

* feat(... ReservationRepository, WaitingRepository): 날짜, 시간, 테마, 멤버가 같은 예약과 예약 대기가 존재하는지 확인하는 메서드 추가

* feat(user-reservation.js): 예약-예약대기, 예약대기-예약대기 일 경우 구체적인 에러 메세지가 alert 창에 뜨도록 수정

* feat(WaitingExceptionCode): 같은 방탈출을 예약 후 예약 대기, 예약 대기 후 예약 대기할 경우에 대한 에러 메세지 추가

* refactor(... ReservationService): jpa 메서드명 변경에 따른 수정

* style(ReservationExceptionCode): 주석 삭제

* docs(README): 예외처리 기능 완료

* style(... RegistrationInfo): 사용하지 않는 생성자, 공백 제거

* refactor(... Waiting): 필드 순서, 생성자 파라미터 순서 변경

* feat(WaitingController): 예약대기 컨트롤러 테스트 추가

* docs(README): 4단계 요구사항 추가

* feat(waiting.html, waiting.js): 예약 대기 관리를 위한 리소스 파일 추가

* feat(AdminController, AdminViewController): 예약 대기 조회 기능 추가

* feat(data.sql): 예약 대기 데이터 추가

* feat(data.sql): 예약 대기 데이터 하나 추가

* feat(... AdminController): 예약 대기 거절 기능 추가

* style(RegistrationInfo): 공백 추가

* feat(AdminController): 예약 대기 승인 api 추가

* rename: waiting, reservation 패키지를 registration 패키지 안에서 domain으로 묶음

* feat(WaitingExceptionCode): 예약 id가 없을 경우 예외 상수 추가

* refactor(waiting.html): 내 예약을 확인하는 api endpoint 변경

* feat(RegistrationService): 예약 대기 승인 서비스 추가

* refactor(ReservationService): 예약 날짜, 시간이 현재 날짜보다 이전인지 확인하는 예외처리를 도메인 -> 서비스로 이동

* style(ReservationFilterRequest): 클래스와 메서드 사이에 공백 추가

* feat(waiting.js): 예약 대기 승인 api 호출을 위한 엔드포인트, http 메서드 종류 추가

* docs(README): 예약 대기 승인 기능 완성

* chore(data.sql): 멤버 이름 구체적으로 변경

* feat(WaitingService): 에약 대기 순번을 반환하는 메서드 추가

* feat(RegistrationService): 예약 대기 승인 메서드 추가, 예약 대기 승인시 발생할 수 있는 예외처리 완료

* feat(ReservationExceptionCode): 예약을 취소하지 않았는데 예약 대기 승인할 경우 예외 추가

* feat(ReservationRepository): 날짜, 테마, 예약 시간이 같은 에약 존재 여부를 반환하는 메서드 추가

* feat(WaitingExceptionCode): 첫번째가 아닌 예약 대기를 승인할 경우 예외 추가

* feat(WaitingRepository): 날짜, 테마id, 시간id로 Waiting을 찾는 메서드, 특정 예약 대기의 대기 순번을 반환하는 메서드 추가

* feat(RegistrationRequest): 예약, 예약 대기의 방탈출 정보를 전달하기 위한 dto 생성

* feat(waiting.js): 예외 발생시 alert 창으로 보이도록 기능 추가

* docs(README): 예약 대기 승인 예외처리 기능 완료

* style(... MemberController): 주석 정리

* refactor(ReservationServiceTest): 예약 날짜 예외처리를 domain -> service 로 옮김에 따라 테스트 메서드도 이동

* docs(README): 새로운 ERD로 수정

* docs(README): 새로운 ERD로 수정

* refactor(Waiting): Waiting이 Reservation id를 갖고있도록 수정

* refactor(... WaitingResponse): Waiting 필드 변화에 따른 수정

* refactor(... MemberService): 예약 대기 자동 승인으로 변경

* feat(MemberRepository): 멤버 save 레포지토리 메서드 추가

* chore(data.sql): waiting 테이블 변화에 따른 데이터 수정

* chore(init.sql): 테스트 메서드 격리를 위한 init sql 추가

* rename(... RegistrationDto): 애플리케이션에서 사용되는 dto는 ...Dto 로 클래스명 변경

* refactor(ReservationService): 매개변수로 dto 사용

* refactor(ReservationController): dto 활용

* feat(WaitingDto): 예약 대기 정보 전달용 dto 생성

* refactor(WaitingService): countWithRank 매개변수 타입 WaitingDto으로 수정

* test(... ReservationTimeServiceTest): 특정 시간대에만 예약 가능한 비즈니스 로직에 맞게 테스트 데이터 수정

* feat(...RegistrationService): 서비스 레이어에 @Transactional 추가

* rename(... RegistrationDto): registration 패키지 삭제

* refactor(WaitingControllerTest): 목 대신 직접 db에 접근해서 예약 대기 테스트

+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -1,80 +1,90 @@
 package roomescape.reservation.service;
 
 import java.time.LocalDate;
+import java.time.LocalTime;
 import java.util.List;
+import java.util.Optional;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import roomescape.admin.domain.FilterInfo;
 import roomescape.admin.dto.AdminReservationRequest;
 import roomescape.admin.dto.ReservationFilterRequest;
-import roomescape.global.exception.model.RoomEscapeException;
+import roomescape.exception.RoomEscapeException;
+import roomescape.exception.model.MemberExceptionCode;
+import roomescape.exception.model.ReservationExceptionCode;
+import roomescape.exception.model.ReservationTimeExceptionCode;
+import roomescape.exception.model.ThemeExceptionCode;
 import roomescape.member.domain.Member;
-import roomescape.member.exception.MemberExceptionCode;
 import roomescape.member.repository.MemberRepository;
 import roomescape.reservation.domain.Reservation;
-import roomescape.reservation.dto.ReservationRequest;
 import roomescape.reservation.dto.ReservationResponse;
 import roomescape.reservation.dto.ReservationTimeAvailabilityResponse;
-import roomescape.reservation.dto.ReservationWaitingResponse;
 import roomescape.reservation.repository.ReservationRepository;
-import roomescape.theme.domain.Theme;
-import roomescape.theme.exception.ThemeExceptionCode;
-import roomescape.theme.repository.ThemeRepository;
+import roomescape.waiting.domain.Waiting;
+import roomescape.waiting.repository.WaitingRepository;
+import roomescape.reservation.dto.RegistrationDto;
 import roomescape.reservationtime.domain.ReservationTime;
-import roomescape.reservationtime.exception.TimeExceptionCode;
-import roomescape.reservationtime.repository.TimeRepository;
+import roomescape.reservationtime.repository.ReservationTimeRepository;
+import roomescape.theme.domain.Theme;
+import roomescape.theme.repository.ThemeRepository;
 
+@Transactional
 @Service
 public class ReservationService {
 
     private final ReservationRepository reservationRepository;
-    private final TimeRepository timeRepository;
+    private final ReservationTimeRepository reservationTimeRepository;
     private final ThemeRepository themeRepository;
     private final MemberRepository memberRepository;
+    private final WaitingRepository waitingRepository;
 
-    public ReservationService(ReservationRepository reservationRepository, TimeRepository timeRepository,
-                              ThemeRepository themeRepository, MemberRepository memberRepository) {
+    public ReservationService(ReservationRepository reservationRepository,
+                              ReservationTimeRepository reservationTimeRepository, ThemeRepository themeRepository,
+                              MemberRepository memberRepository, WaitingRepository waitingRepository) {
         this.reservationRepository = reservationRepository;
-        this.timeRepository = timeRepository;
+        this.reservationTimeRepository = reservationTimeRepository;
         this.themeRepository = themeRepository;
         this.memberRepository = memberRepository;
+        this.waitingRepository = waitingRepository;
     }
 
-    public ReservationResponse addReservation(ReservationRequest reservationRequest, long memberId) {
-        ReservationTime time = timeRepository.findById(reservationRequest.timeId())
-                .orElseThrow(() -> new RoomEscapeException(TimeExceptionCode.FOUND_TIME_IS_NULL_EXCEPTION));
-        Theme theme = themeRepository.findById(reservationRequest.themeId())
+    public ReservationResponse addReservation(RegistrationDto registrationDto) {
+        ReservationTime time = reservationTimeRepository.findById(registrationDto.timeId())
+                .orElseThrow(() -> new RoomEscapeException(ReservationTimeExceptionCode.FOUND_TIME_IS_NULL_EXCEPTION));
+        Theme theme = themeRepository.findById(registrationDto.themeId())
                 .orElseThrow(() -> new RoomEscapeException(ThemeExceptionCode.FOUND_THEME_IS_NULL_EXCEPTION));
-        Member member = memberRepository.findMemberById(memberId)
+        Member member = memberRepository.findMemberById(registrationDto.memberId())
                 .orElseThrow(() -> new RoomEscapeException(ThemeExceptionCode.FOUND_MEMBER_IS_NULL_EXCEPTION));
 
-        Reservation saveReservation = new Reservation(reservationRequest.date(), time, theme, member);
+        validateDateAndTimeWhenSave(registrationDto.date(), time);
+        Reservation saveReservation = new Reservation(registrationDto.date(), time, theme, member);
 
-        return ReservationResponse.fromReservation(reservationRepository.save(saveReservation));
+        return ReservationResponse.from(reservationRepository.save(saveReservation));
     }
 
     public void addAdminReservation(AdminReservationRequest adminReservationRequest) {
-        ReservationTime time = timeRepository.findById(adminReservationRequest.timeId())
-                .orElseThrow(() -> new RoomEscapeException(TimeExceptionCode.FOUND_TIME_IS_NULL_EXCEPTION));
+        ReservationTime time = reservationTimeRepository.findById(adminReservationRequest.timeId())
+                .orElseThrow(() -> new RoomEscapeException(ReservationTimeExceptionCode.FOUND_TIME_IS_NULL_EXCEPTION));
         Theme theme = themeRepository.findById(adminReservationRequest.themeId())
                 .orElseThrow(() -> new RoomEscapeException(ThemeExceptionCode.FOUND_THEME_IS_NULL_EXCEPTION));
         Member member = memberRepository.findMemberById(adminReservationRequest.memberId())
                 .orElseThrow(() -> new RoomEscapeException(MemberExceptionCode.MEMBER_NOT_EXIST_EXCEPTION));
 
+        validateDateAndTimeWhenSave(adminReservationRequest.date(), time);
         Reservation saveReservation = new Reservation(adminReservationRequest.date(), time, theme, member);
-        ReservationResponse.fromReservation(reservationRepository.save(saveReservation));
+        ReservationResponse.from(reservationRepository.save(saveReservation));
     }
 
-
     public List<ReservationResponse> findReservations() {
-        List<Reservation> reservations = reservationRepository.findAllByOrderByDateAscTimeAsc();
+        List<Reservation> reservations = reservationRepository.findAllByOrderByDateAscReservationTimeAsc();
 
         return reservations.stream()
-                .map(ReservationResponse::fromReservation)
+                .map(ReservationResponse::from)
                 .toList();
     }
 
     public List<ReservationTimeAvailabilityResponse> findTimeAvailability(long themeId, LocalDate date) {
-        List<ReservationTime> allTimes = timeRepository.findAllByOrderByStartAt();
+        List<ReservationTime> allTimes = reservationTimeRepository.findAllByOrderByStartAt();
         List<Reservation> reservations = reservationRepository.findAllByThemeIdAndDate(themeId, date);
         List<ReservationTime> bookedTimes = extractReservationTimes(reservations);
 
@@ -88,20 +98,33 @@ public class ReservationService {
 
         return reservationRepository.findAllByMemberIdAndThemeIdAndDateBetween(filterInfo.getMemberId(),
                         filterInfo.getThemeId(), filterInfo.getFromDate(), filterInfo.getToDate()).stream()
-                .map(ReservationResponse::fromReservation)
+                .map(ReservationResponse::from)
                 .toList();
     }
 
-    public List<ReservationWaitingResponse> findMemberReservations(long id) {
+    public List<ReservationResponse> findMemberReservations(long id) {
         List<Reservation> reservations = reservationRepository.findAllByMemberId(id);
 
         return reservations.stream()
-                .map(ReservationWaitingResponse::from)
+                .map(ReservationResponse::from)
                 .toList();
     }
 
-    public void removeReservations(long reservationId) {
-        reservationRepository.deleteById(reservationId);
+    @Transactional
+    public void removeReservation(long reservationId) {
+        Optional<Waiting> waiting = waitingRepository.findFirstByReservationIdOrderByCreatedAt(reservationId);
+
+        if (waiting.isEmpty()) {
+            reservationRepository.deleteById(reservationId);
+            return;
+        }
+
+        Reservation reservation = reservationRepository.findById(reservationId)
+                .orElseThrow(() -> new RoomEscapeException(ReservationExceptionCode.RESERVATION_NOT_EXIST));
+        reservation.setMember(waiting.get().getMember());
+
+        reservationRepository.save(reservation);
+        waitingRepository.deleteById(waiting.get().getId());
     }
 
     private List<ReservationTime> extractReservationTimes(List<Reservation> reservations) {
@@ -112,5 +135,21 @@ public class ReservationService {
 
     private boolean isTimeBooked(ReservationTime time, List<ReservationTime> bookedTimes) {
         return bookedTimes.contains(time);
+    }
+
+    private void validateDateAndTimeWhenSave(LocalDate date, ReservationTime time) {
+        if (date.isBefore(LocalDate.now())) {
+            throw new RoomEscapeException(ReservationExceptionCode.RESERVATION_DATE_IS_PAST_EXCEPTION);
+        }
+
+        if (date.equals(LocalDate.now())) {
+            validateTime(time);
+        }
+    }
+
+    private void validateTime(ReservationTime time) {
+        if (time.isBeforeTime(LocalTime.now())) {
+            throw new RoomEscapeException(ReservationExceptionCode.RESERVATION_TIME_IS_PAST_EXCEPTION);
+        }
     }
 }
